Move Header component from top level App
Header component was moved to from App component to each of the each
route in the website. The reason for this is because since the hidden
folks will be displayed in the Header component when you start playing
the game, those folks will still be there even when you navigate to
other route.

Since that is not the design I wanted I had to make it be explixitly
called in all routes then made the hidenFolks props optional which will
only be passed where it's needed.